requirements.txt
flask
spacy
python-docx
PyPDF2
